staking: Moves all StakingLedger logic to staking/src/ledger.rs
